- Tried to reduce the number of compiler warnings generated.
- The cross compiler GNUmakefile.os4 variant now also logs all warnings
  and errors to a compiler.log file.

+++ math_lround.c
@@ -82,7 +82,7 @@ lround(double x)
           result = msw >> (20 - exponent_less_1023);
         }
     }
-  else if (exponent_less_1023 < (8 * sizeof (long int)) - 1)
+  else if (exponent_less_1023 < (LONG)(8 * sizeof (long int)) - 1)
     {
       if (exponent_less_1023 >= 52)
         result = ((long int) msw << (exponent_less_1023 - 20)) | (lsw << (exponent_less_1023 - 52));
